Major fix: diag extension, prevent overlap on seq B
There was only one check, made on overlaps for
sequence A (overlapping columns)
This commit adds a check for overlaps on rows.
Without it, it was possible that some diagonals would extend in excess
if sequence A was containing repeats that were also
present at the end of the secgment B of the previous diagonal.