fix: range error reported in estring difference
This happened for a special case of estring termination where the
estringA is entirely parsed and
there is just one negative entry from B remaining.
In that case, lValB is descrease by the value from A which is undefined,
resulting in lValB being NaN.
This fix adds a check on lValA: if it is undefined than lValB is not changed
and the last iteration can occur.