transition DEL->MATCH might use wrong del value
Previous code was getting the value for del from the del array, but it's possible that this value has been updated by the computation of the optimal delete transition in the same loop.
Instead, we create an intermediary variable to store the original value and use it to compute the score for the transition from del to match.